% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gillpack-package.R
\docType{package}
\name{gillpack-package}
\alias{gillpack}
\alias{gillpack-package}
\title{gillpack: Packed to the gills with helper functions by (and mostly for) Amy Gill}
\description{
Helper functions to make it easier to work with biochemical reaction network simulations and their outputs from BNGL or MATLAB.
}
\author{
\strong{Maintainer}: Amy Gill \email{gill.signals@gmail.com}

}
\keyword{internal}
