% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cb_pal.R
\name{load_cb_pal}
\alias{load_cb_pal}
\title{Load a colorblind-friendly palette (from Wong 2011)}
\usage{
load_cb_pal(list = TRUE)
}
\arguments{
\item{list}{Boolean determining whether output is a named list (TRUE, default) or a named vector (FALSE)}
}
\value{
if \code{list=TRUE}: named list with values "black", "red", "orange", "green", "yellow", "darkblue", "pink";
if \code{list=FALSE}: named vector with same values
}
\description{
Load a colorblind-friendly palette (from Wong 2011)
}
\examples{
cb_pal <- load_cb_pal()
cb_pal$red  # or cb_pal[[2]]

cb_pal_vec <- load_cb_pal(list = FALSE)
cb_pal_vec["red"] # or cb_pal_vec[2]
}
