% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rxn_stoich_mat.R
\name{make_rxn_stoich_mat}
\alias{make_rxn_stoich_mat}
\title{Make reaction stoichiometry matrix from reactant and product lists}
\usage{
make_rxn_stoich_mat(reactant_list, product_list, species_names)
}
\arguments{
\item{reactant_list}{A list of length n_reactions; list elements are
character vectors of the index of each species consumed by this reaction;
the number of times each reaction occurs matches reaction stoichiometry}

\item{product_list}{A list of length n_reactions; list elements are
character vectors of the index of each species produced by this reaction;
the number of times each product occurs matches reaction stoichiometry}

\item{species_names}{A character vector of length n_species mapping index numbers to species}
}
\value{
matrix with dimensions (rows = n_rxns, cols = n_species);
sparse (mostly 0) with +- 1 (or other low integers) encoding reactions;
for each reaction row, reactants are negative and products positive;
value = net number of that species created/destroyed by that reaction
}
\description{
Make reaction stoichiometry matrix from reactant and product lists
}
\examples{
print("In development...")

}
