% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_net_file.R
\name{read_net_file}
\alias{read_net_file}
\title{Title}
\usage{
read_net_file(
  filename,
  col_names = c("index", "name", "value", "remove", "const_expr")
)
}
\arguments{
\item{filename}{Output file name/path}

\item{col_names}{Vector of column names for imported table; defaults provided}
}
\value{
a list with elements named with \code{col_names}
}
\description{
Title
}
\examples{
print("In development...")
}
