% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{string to split (character vector with one element)}

\item{split}{delimiter to split on (character vector with one element)}
}
\value{
a character vector with elements matching the original string split by the delimiter
}
\description{
Split a string
}
