% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_param_df.R
\name{update_param_df}
\alias{update_param_df}
\title{Modify a parameter data frame with calculated parameter values}
\usage{
update_param_df(params)
}
\arguments{
\item{params}{data frame extracted from list}
}
\value{
modified parameter data frame
}
\description{
Modify a parameter data frame with calculated parameter values
}
\examples{
print("In development...")
}
